This has all internal PEXes be built with settings to improve performance:

- with `--no-pre-install-wheels`, to package `.whl` directly rather than unpack and install them. (NB. this requires Pex 2.3.0 to pick up https://github.com/pex-tool/pex/pull/2392)
- with `PEX_MAX_INSTALL_JOBS`, to use more concurrency for install, when available

This is designed to be a performance improvement for any processing where Pants synthesises a PEX internally, like `pants run path/to/script.py` or `pants test ...`. https://github.com/pex-tool/pex/issues/2292 has benchmarks for the PEX tool itself.

For benchmarks, I did some more purposeful ones with tensorflow (PyTorch seems a bit awkward hard to set-up and Tensorflow is still huge), using https://gist.github.com/huonw/0560f5aaa34630b68bfb7e0995e99285 . I did 3 runs each of two goals, with 2.21.0.dev4 and with `PANTS_SOURCE` pointing to this PR, and pulled the numbers out by finding the relevant log lines:

- `pants --no-local-cache --no-pantsd --named-caches-dir=$(mktemp -d)  test example_test.py`. This involves building 4 separate PEXes partially in parallel, partially sequentially: `requirements.pex`, `local_dists.pex` `pytest.pex`, and then `pytest_runner.pex`. The first and last are the interesting ones for this test.
- `pants --no-local-cache --no-pantsd --named-caches-dir=$(mktemp -d) run script.py`. This just builds the requirements into `script.pex`.

(NB. these are potentially unrealistic in they're running with all caching turned off or cleared, so are truly a worst case. This means they're downloading tensorflow wheels and all the others, each time, which takes about 30s on my 100Mbit/s connection. Faster connections will thus see a higher ratio of benefit.)

| goal | period  | before (s) | after (s) |
|---|---|---:|---:|
| `run script.py`  | building requirements | 74-82 | 49-52 |
| `test some_test.py` | building requirements | 67-71 | 30-36 |
| | building pytest runner  | 8-9 | 17-18 |
| | total to start running tests | 76-80 | 53-58 |
 
I also did more adhoc ones on a real-world work repo of mine, which doesn't use any of the big ML libraries, just running some basic goals once.

| goal | period  | before (s) | after (s) |
|---|---|---:|---:|
| `pants export` on largest resolve |  building requirements | 66 | 35 |
| | total  | 82 | 54 |
| "random" `pants test path/to/file.py` (1 attempt) | building requirements and pytest runner | 1 | 49 | 38 |

Two explicit questions for review:

1. The `--no-pre-install-wheels` flag behaviour isn't explicitly tested... but maybe it should be, i.e. validate we're passing this flag for internal PEXes. Thoughts?
2. Setting `PEX_MAX_INSTALL_JOBS` as an env var may result in fewer cache hits, e.g. `pants test path/to/file.py` may not be able to reuse caches from `pants test ::` (and vice versa) because the available concurrency is different, and thus this may be better to do differently. Thoughts?
   - for instance, leave it as it was (the default `1`)
   - set to `0` to have each PEX process do its own concurrency, e.g. based on number of CPU cores and wheels. Multiple processes in parallel risk overloading the machine since they don't coordinate, though.

Fixes #15062