@huonw 2024-03-24
It looks like the local dists test failure is https://github.com/pex-tool/pex/issues/2391, so this is waiting on that.